+++ numina/core/tests/test_recipe.py
@@ -16,7 +16,7 @@ from ..metarecipes import RecipeType
 from ..recipes import BaseRecipe
 from ..recipeinout import RecipeInput, RecipeResult
 from ..requirements import ObservationResultRequirement
-from ..dataholders import Product
+from ..dataholders import Result
 from ..requirements import Requirement
 from numina.types.obsresult import QualityControlProduct
 from numina.types.qc import QC
@@ -24,7 +24,7 @@ from numina.types.qc import QC
 
 class PruebaRecipe1(BaseRecipe):
     somereq = Requirement(int, 'Some integer')
-    someresult = Product(int, 'Some integer')
+    someresult = Result(int, 'Some integer')
 
     def run(self, recipe_input):
         if recipe_input.somereq >= 100:
@@ -65,7 +65,7 @@ def test_metaclass():
 
     class RecipeTest(with_metaclass(RecipeType, object)):
         obsresult = ObservationResultRequirement()
-        someresult = Product(int, 'Some integer')
+        someresult = Result(int, 'Some integer')
 
     assert hasattr(RecipeTest, 'RecipeInput')
 
@@ -79,11 +79,11 @@ def test_metaclass():
 def test_recipe_with_autoqc():
 
     class RecipeTestAutoQC(BaseRecipe):
-        qc82h = Product(QualityControlProduct, destination='qc')
+        qc82h = Result(QualityControlProduct, destination='qc')
 
     class RecipeTest(RecipeTestAutoQC):
         obsresult = ObservationResultRequirement()
-        someresult = Product(int, 'Some integer')
+        someresult = Result(int, 'Some integer')
 
     assert hasattr(RecipeTest, 'RecipeInput')
 
@@ -101,7 +101,7 @@ def test_recipe_with_autoqc():
     assert 'qc' in RecipeTest.products()
 
     for prod in RecipeTest.RecipeResult.stored().values():
-        assert isinstance(prod, Product)
+        assert isinstance(prod, Result)
 
     qc = RecipeTest.RecipeResult.stored()['qc']
 
@@ -112,7 +112,7 @@ def test_recipe_without_autoqc():
 
     class RecipeTest(BaseRecipe):
         obsresult = ObservationResultRequirement()
-        someresult = Product(int, 'Some integer')
+        someresult = Result(int, 'Some integer')
 
     assert hasattr(RecipeTest, 'RecipeInput')
 
@@ -130,18 +130,18 @@ def test_recipe_without_autoqc():
     assert 'qc' not in RecipeTest.products()
 
     for prod in RecipeTest.RecipeResult.stored().values():
-        assert isinstance(prod, Product)
+        assert isinstance(prod, Result)
 
 
 def test_recipe_io_inheritance():
 
     class TestBaseRecipe(BaseRecipe):
         obresult = ObservationResultRequirement()
-        someresult1 = Product(int, 'Some integer')
+        someresult1 = Result(int, 'Some integer')
 
     class RecipeTest(TestBaseRecipe):
         other = Requirement(int, description='Other')
-        someresult2 = Product(int, 'Some integer')
+        someresult2 = Result(int, 'Some integer')
 
     assert issubclass(RecipeTest.RecipeInput, TestBaseRecipe.RecipeInput)
 
@@ -174,7 +174,7 @@ def test_recipe_io_baseclass():
         RecipeResult = MyRecipeResult
 
         other = Requirement(int, description='Other')
-        someresult2 = Product(int, 'Some integer')
+        someresult2 = Result(int, 'Some integer')
 
     assert issubclass(RecipeTest.RecipeInput, MyRecipeInput)